refactor: fix and reorganise import statements
