+++ frontend/index.html
@@ -4,12 +4,13 @@
     <meta charset="UTF-8" />
     <meta http-equiv="content-type" content="content/html" />
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
-    <meta name="viewport" content="width=device-width, account-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0" />
     <meta name="language" content="en" />
-    <meta name="robots" content="noindex,nofollow" />
-    <meta name="description" content="" />
+    <meta name="robots" content="noindex, nofollow" />
+    <meta name="description" content="university application management tool" />
+    <meta name="keywords" content="university, application, student, mentor">
 
-    <title>Daigaku</title>
+    <title>Daigaku 大学</title>
 
     <link rel="author" href="https://linkedin.com/in/tamasnvk/" />
     <link href="https://fonts.googleapis.com/css2?family=Raleway:wght@100;400;600;800&display=swap" rel="stylesheet" />
