+++ frontend/src/components/page/home/ForgottenPasswordForm/ForgottenPasswordForm.hooks.tsx
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { UseFormSetError } from 'react-hook-form';
-import accountService from '@services/account/Account.service.ts';
+import accountService from '@services/account/account.service.ts';
 import { mutationKeys } from '@configuration';
 import { ConfirmationModalT } from '@pages/shared/Home/Home.types.ts';
