+++ frontend/src/components/page/new-application/NewApplicationForm/NewApplicationForm.component.tsx
@@ -17,9 +17,9 @@ import {
   SelectCountry,
   SelectUniversity,
 } from '@components/shared/field-implementations';
-import { Toast } from 'components/shared/notification';
-import { CountryOptionT } from '@services/country/Country.service.ts';
-import { UniversityOptionT } from '@services/university/University.service.ts';
+import { Toast } from '@components/shared/notification';
+import { CountryOptionT } from '@services/country/country.service.ts';
+import { UniversityOptionT } from '@services/university/university.service.ts';
 import {
   countryInformation,
   formInformation,
