--- frontend/src/components/shared/field-implementations/GeneralSelectInputField/index.ts
@@ -1,3 +1,0 @@
-import GeneralSelectInputField from './GeneralSelectInputField.component.tsx';
-
-export default GeneralSelectInputField;
