+++ frontend/src/components/shared/field-implementations/GenericSelectInputField/index.ts
@@ -1,0 +1,3 @@
+import GenericSelectInputField from './GenericSelectInputField.component.tsx';
+
+export default GenericSelectInputField;
