+++ frontend/src/components/shared/field-implementations/PasswordInputField/PasswordInputField.hooks.tsx
@@ -13,4 +13,6 @@ const useRevealPassword = () => {
   };
 };
 
-export default useRevealPassword;
+export {
+  useRevealPassword,
+};
