+++ frontend/src/hooks/country/index.ts
@@ -1,0 +1,5 @@
+import useGetCountryOptions from './useGetCountryOptions.tsx';
+
+export {
+  useGetCountryOptions,
+};
