+++ frontend/src/hooks/university/index.ts
@@ -1,0 +1,5 @@
+import useGetUniversityOptionsByCountryUuid from './useGetUniversityOptionsByCountryUuid.tsx';
+
+export {
+  useGetUniversityOptionsByCountryUuid,
+};
