+++ frontend/src/pages/shared/Applications/Applications.hooks.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { QueryClient } from '@tanstack/react-query';
 import { queryKeys } from '@configuration';
-import { ApplicationT } from '@custom-types/ApplicationT.ts';
+import { ApplicationT } from '@services/application/application.service.ts';
 
 export type ColumnT = {
   id: string;
