+++ frontend/src/services/application/application.service.ts
@@ -1,9 +1,25 @@
 import { AxiosResponse } from 'axios';
 import { axiosConfigWithAuth } from '@configuration';
-import { ApplicationT } from '@custom-types/ApplicationT.ts';
 import { NewApplicationFormFieldsT } from '@components/page/new-application/NewApplicationForm/NewApplicationForm.hooks.tsx';
 import { DashboardDataT } from '@pages/shared/Dashboard/Dashboard.hooks.tsx';
 import { UpdateApplicationFormFieldsT } from '@components/page/application/ApplicationForm/ApplicationForm.hooks.tsx';
+
+export type ApplicationT = {
+  uuid: string;
+  account: string;
+  country: string;
+  university: string;
+  courseName: string;
+  minorSubject: string;
+  programmeLength: number;
+  applicationStatus: string;
+  interviewStatus: string;
+  offerStatus: string;
+  responseStatus: string;
+  finalDestinationStatus: string;
+  createdAt: Date;
+  lastUpdatedAt: Date;
+}
 
 const applicationService = {
   getByUuid: async (applicationUuid: string): Promise<AxiosResponse<ApplicationT>> => {
