+++ frontend/src/services/application/applicationStatus.service.ts
