+++ frontend/src/services/application/offerStatus.service.ts
