--- frontend/src/types/ApplicationT.ts
@@ -1,16 +1,0 @@
-export type ApplicationT = {
-  uuid: string;
-  account: string;
-  country: string;
-  university: string;
-  courseName: string;
-  minorSubject: string;
-  programmeLength: number;
-  applicationStatus: string;
-  interviewStatus: string;
-  offerStatus: string;
-  responseStatus: string;
-  finalDestinationStatus: string;
-  createdAt: Date;
-  lastUpdatedAt: Date;
-}
